Fixed loss of precision when dividing integers Closes https://github.com/SoftUni-Internal/suls-issues/issues/4893

+++ Open Judge System/Workers/OJS.Workers.LocalWorker/SubmissionJobStrategy.cs
@@ -145,10 +145,10 @@
                 }
                 else
                 {
-                    var coefficient = this.submission.CorrectTestRunsWithoutTrialTestsCount /
-                                      this.submission.TestsWithoutTrialTestsCount;
+                    var coefficient = (double)this.submission.CorrectTestRunsWithoutTrialTestsCount /
+                        this.submission.TestsWithoutTrialTestsCount;
 
-                    this.submission.Points = coefficient * this.submission.Problem.MaximumPoints;
+                    this.submission.Points = (int)(coefficient * this.submission.Problem.MaximumPoints);
                 }
             }
             catch (Exception ex)
